requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==5.1.1
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.2
gspread-pandas==3.2.0
idna==3.3
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
